A more efficient implementation for large sparse bitmaps.
We store the data in 6 levels, each level being a bitmap that
indicates if any bits in the slots below are set. This gives a
massive performance improvement for large sparse bitmaps, but
doesn't perform that well on dense bitmaps. We use some tricks
to make the dense bitmaps suck less.