The jump table should be static.
For some reason the jump table wasn't static. I could swear I made
it static when I wrote the code the first time.

The difference is that clang pre-computed the jump table even though
it wasn't static while gcc didn't, it generated the jump table on
every call to the function, making it very slow.